[BUH-5] Add event disbanding
https://buhinder.atlassian.net/browse/BUH-5
